Fix #95 and add an optional admin SQL query interface

+++ src/admin/analytics/index.rs
@@ -135,7 +135,7 @@ struct SummariesDuration {
 struct Summary {
     #[serde(with = "http_serde::uri")]
     uri: Uri,
-    visits: u32,
+    visits: u64,
 }
 
 #[derive(Serialize)]
